refactor(Services): wrap Formatters non-Sample + ReadCommands non-Sample types under Services namespace; rename files
Continues the Services.Formatters wrap series started in #376. Wraps 7
more public types in Services/Formatters/ under Services.Formatters,
plus 6 public types in Services/ReadCommands/ under Services. All
caller sweeps + file renames included so the rule
'<Namespace>+<Type>.swift' applies uniformly across the Services
subtargets.

## Renames (Services.Formatters.*)

| Before | After |
|---|---|
| JSONSearchResultFormatter   | Services.Formatters.JSONSearchResultFormatter |
| FrameworksJSONFormatter     | Services.Formatters.FrameworksJSONFormatter |
| TextSearchResultFormatter   | Services.Formatters.TextSearchResultFormatter |
| FrameworksTextFormatter     | Services.Formatters.FrameworksTextFormatter |
| HIGTextFormatter            | Services.Formatters.HIGTextFormatter |
| HIGJSONFormatter            | Services.Formatters.HIGJSONFormatter |
| TeaserResults               | Services.Formatters.TeaserResults |

## Renames (Services.*)

| Before | After |
|---|---|
| DocsSearchService           | Services.DocsSearchService |
| HIGQuery                    | Services.HIGQuery |
| HIGSearchService            | Services.HIGSearchService |
| ReadService                 | Services.ReadService |
| TeaserService               | Services.TeaserService |
| UnifiedSearchService        | Services.UnifiedSearchService |

## File renames

Services/Formatters/
- JSONFormatter.swift           -> Services.Formatters+JSONSearchResultFormatter.swift
                                  (dominant: JSONSearchResultFormatter;
                                  FrameworksJSONFormatter companion)
- TextFormatter.swift           -> Services.Formatters+TextSearchResultFormatter.swift
                                  (dominant: TextSearchResultFormatter;
                                  FrameworksTextFormatter companion)
- HIGFormatters.swift           -> Services.Formatters+HIGTextFormatter.swift
                                  (dominant: HIGTextFormatter;
                                  HIGJSONFormatter companion)
- TeaserResults.swift           -> Services.Formatters+TeaserResults.swift

Services/ReadCommands/
- DocsSearchService.swift       -> Services+DocsSearchService.swift
- HIGSearchService.swift        -> Services+HIGSearchService.swift
                                  (carries HIGQuery companion)
- ReadService.swift             -> Services+ReadService.swift
- TeaserService.swift           -> Services+TeaserService.swift
                                  (carries 'extension Services.ServiceContainer'
                                  helper retargeted to its sibling)
- UnifiedSearchService.swift    -> Services+UnifiedSearchService.swift
                                  (also carries an
                                  'extension Services.ServiceContainer'
                                  helper)

## Mechanics

- Each top-level public type wraps individually in 'extension Services'
  or 'extension Services.Formatters'. File-scope helper extensions
  ('extension <WrappedType> { ... }' / 'extension Services.ServiceContainer
  { ... }') retarget alongside.
- 15 caller files swept: CompositeToolProvider, Search.PackageQuery,
  Services.swift docstring, Services+ServiceContainer.swift,
  Command.Search+SourceRunners, Command+Read, Command+Search,
  MarkdownFormatter (cross-file refs), the now-renamed Formatters /
  ReadCommands files themselves, plus ServicesTests.

## Still-pending (next PR)

The FOOTER formatter family + MarkdownFormatter + UnifiedSearchFormatters
remain unwrapped in this PR — they have heavier file-scope
'public extension SearchFooter { ... }' / multi-type complexity that
needs its own focused PR. User has indicated a target layout
('FooterFormatter <- Formatter.Footer') that implies a deeper
sub-namespace structure with 'Formatter' (singular) and per-family
sub-namespaces ('Formatter.Footer', 'Formatter.JSON', 'Formatter.Markdown',
'Formatter.Text'). That restructure may also rename the existing
'Services.Formatters.*' (plural) to 'Services.Formatter.*' (singular)
to match. Tracked as a follow-up.

## Verification

- xcrun swift build clean.
- xcrun swift test: 1300/1300 passing.

Part of the namespacing sweep tracked in #183. Previous Services-related:
#354, #375, #376.
